# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Marketplace API",
        "description": "Marketplace API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/login": {
            "post": {
                "tags": [
                    "Авторизация"
                ],
                "summary": "Логин",
                "operationId": "bae68d1c55329ecb6e41f48285840c8f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "description": "Email",
                                        "type": "string",
                                        "example": "admin@marketplace.com"
                                    },
                                    "password": {
                                        "description": "Password",
                                        "type": "string",
                                        "example": "Aa123123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Успешно"
                    },
                    "404": {
                        "description": "Пользовтель не найден"
                    },
                    "401": {
                        "description": "Не авторизован"
                    },
                    "422": {
                        "description": "Ошибка проверки"
                    }
                }
            }
        },
        "/api/v1/logout": {
            "post": {
                "tags": [
                    "Авторизация"
                ],
                "summary": "Выход",
                "operationId": "c5cdd87fdbf9d0f95dede875bd023dec",
                "responses": {
                    "200": {
                        "description": "Успешно"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/v1/categories/get": {
            "get": {
                "tags": [
                    "Категории"
                ],
                "summary": "Получить список категорий",
                "operationId": "d79e90761b64470a9b423f358ffc749b",
                "responses": {
                    "200": {
                        "description": "Список категорий",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Category"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/v1/categories/create": {
            "post": {
                "tags": [
                    "Категории"
                ],
                "summary": "Добавить категорию",
                "operationId": "12074baf8dd0ef9699be5823f68a388d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "parent_id": {
                                        "description": "ID родительской категории",
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "name": {
                                        "description": "Название категории",
                                        "type": "string"
                                    },
                                    "slug": {
                                        "description": "Уникальный идентификатор категории",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "Описание категории",
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "status": {
                                        "description": "Статус категории",
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Категория успешно добавлена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Category"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/v1/categories/{id}/update": {
            "put": {
                "tags": [
                    "Категории"
                ],
                "summary": "Обновить категорию",
                "operationId": "dd78c127bcc3e22c68ce3d21b5dafbd1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID категории",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "parent_id": {
                                        "description": "ID родительской категории",
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "name": {
                                        "description": "Название категории",
                                        "type": "string"
                                    },
                                    "slug": {
                                        "description": "Уникальный идентификатор категории",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "Описание категории",
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "status": {
                                        "description": "Статус категории",
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Категория успешно обновлена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Category"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/v1/categories/{id}/delete": {
            "delete": {
                "tags": [
                    "Категории"
                ],
                "summary": "Удалить категорию",
                "operationId": "f8a389e712a4f1eec38e26a97aa1fef9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID категории",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Категория успешно удалена"
                    },
                    "404": {
                        "description": "Категория не найдена"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Category": {
                "title": "Category",
                "description": "Схема категории",
                "properties": {
                    "id": {
                        "description": "ID категории",
                        "type": "integer"
                    },
                    "parent_id": {
                        "description": "ID родительской категории",
                        "type": "integer",
                        "nullable": true
                    },
                    "name": {
                        "description": "Название категории",
                        "type": "string"
                    },
                    "slug": {
                        "description": "Уникальный идентификатор категории",
                        "type": "string"
                    },
                    "description": {
                        "description": "Описание категории",
                        "type": "string",
                        "nullable": true
                    },
                    "status": {
                        "description": "Статус категории (активна или нет)",
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "in": "header",
                "scheme": "bearer"
            }
        }
    },
...
}